Вынес запись в файл в отдельный сервис
